requirements.txt
pytest==7.4.2
requests==2.31.0
jsonschema==4.19.1
allure-pytest==2.13.2
curlify==2.2.1
python-dotenv==1.0.0